+++ languages/javascript/src/jest/invisible-setup.js
@@ -1,0 +1,27 @@
+const {
+  enableInvisibleMode,
+  disableInvisibleMode,
+  patchRequireWithIsolation,
+  restoreOriginalRequire
+} = require('../invisible');
+
+function setupForJest(options = {}) {
+  enableInvisibleMode(options);
+  patchRequireWithIsolation();
+
+  if (typeof afterAll === 'function') {
+    afterAll(() => {
+      teardownForJest();
+    });
+  }
+}
+
+function teardownForJest() {
+  restoreOriginalRequire();
+  disableInvisibleMode();
+}
+
+module.exports = {
+  setupForJest,
+  teardownForJest
+};