+++ packages/webpack-plugin-adaptive/index.js
@@ -1,2 +1,17 @@
-// Temporary shim - package moved to languages/javascript/plugins/webpack-plugin-adaptive
-module.exports = require('../../languages/javascript/plugins/webpack-plugin-adaptive');
+class AdaptiveTestsWebpackPlugin {
+  constructor(options = {}) {
+    this.options = options;
+  }
+
+  apply(compiler) {
+    compiler.hooks.beforeRun.tap('AdaptiveTestsWebpackPlugin', () => {
+      if (!this.options.quiet) {
+        // eslint-disable-next-line no-console
+        console.warn('Adaptive Tests Webpack plugin (preview) enabled – run adaptive tests via your test runner.');
+      }
+    });
+  }
+}
+
+module.exports = AdaptiveTestsWebpackPlugin;
+module.exports.default = AdaptiveTestsWebpackPlugin;